at machine Dell at network NetVirtua 10Mbps at Santos, SP - Brazil, at laptop Dell XPS L502X processor i7 2670QM 8GB RAM 1 TB Disk 1.5TB Backup Samsung Disk at Santos SP, Brazil

+++ os/win_window.cpp
@@ -6649,7 +6649,7 @@ namespace win
 
       {
 
-         memset(m_pcolorref, 0, m_size.area() * 4);
+         memset(m_pcolorref, 0, (size_t) (m_size.area() * 4));
 
          _001Print(m_spg);
